Update localhost URL in test environment
The localhost URL in the list of allowed origins has been updated specifically for the test environment. The previous code was allowing localhost in both development and test environments, whereas the updated code is ensuring the localhost:3000 url is only pushed to the allowed origins list in the test environment.